Actually start git-gui
The mount points will only be set up in MSys when the shell is invoked
with --login.  Unfortunately, this means that a console window pops
up briefly.  Maybe we succeed in suppresing that later.

Next step: extract the current directory from info->lpDirectoryW and
cd to it before executing git gui, i.e. something like

	"%s\\bin\\sh.exe --login \"cd %s; git gui\"", msysPath, directory

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>