Implement self-registration of the extension
This is an implementation of TODO "Need to not hard-code the dll path"

Registration is implemented with a minimal regedit engine: an array of
registry values { path, name, value } is the data source. So, all
registry entries are hard-coded in dll.c. The engine attempts to
convert values into longs to create REG_DWORD values.

msysGit (for PathToMsys) is searched in the following order:
- $(TARGET)/..
- $(TARGET)/../..
- %PATH%
- InstallLocation of uninstall info (registration-type dependant).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>