Preview implementation of "smart" context menu
As Johannes Schindelin suggested, it verifies that the selected item is
a valid Git repository and add different menu items accordingly.
This version puts the basic infrastructure in place, and does not
necessarily provide a full-grown menu structure yet. It implements the
suggestion in a way that the text of the item is context-sensitive, but
the action is always the same.

TODO: think about making LONGEST_MENU_ITEM and MAX_PROCESSING_TIME
      dynamic/configurable.
TODO: decide on a structure and actions of added menu items.