Add ability to use PATH from the specified environment
If mingw_spawnvpe is called with a specified environment, we want to use
the PATH variable from the new environment, because it may be adjusted
to include git.exe or sh.exe. Thus, we need to pass the last parameter
of mingw_spawnvpe to all calls of get_path_split. If PATH variable is not
found in the specified environment or the environment was not specified at
all, we'll use PATH from the current one, as before.