Separate struct git_data from Windows-specific definitions
In preparation for the new, platform-independent menu engine, the commonly
used git_data structure is now in a separate file - menuengine.h. Note that
this file is supposed to be platform-neutral.

To avoid nested includes, every file that depends on the old location of
the structure definition now depends on the new file as well.

Signed-off-by: Christian MICHON <christian.michon@gmail.com>