Add support to add a checkmark to a menu item
Currently, the support is limited only to custom item builders (i.e. no
field in menu_item can be used to put the checkmark). The platform is
responsible to mark the menu item in the platform-specific way (on Windows
a small checkmark will appear to the left of item's text).