Add Branch sub-menu
The result of git branch is displayed in the sub-menu. The current branch
is checked. Selecting the item will perform git checkout and notify about
the result.

The number of items is limited by the platform, so, for example, OpenOffice
repo may not have all the branches in the menu. Unfortunately, there is
currently no way to indicate that not all branches are displayed.