Rename 64bit shell extension to allow both to exist.
It would seem to make sense to allow both the 32bit and 64bit
shell extension under windows to exist at the same time, so
I've renamed in the 64bit environment from git_shell_ext.dll to
git_shell_ext64.dll.

Also added a platform-clean target which is just like clean, but
only deletes the dll for the context for which make is being
called.

Signed-off-by: Peter Kahle <pkahle@kahlilia.org>