Support registry entries specific to some versions of Windows.
The handling of the various registry classes is different between Windows
XP and Windows 7. On XP we need to register the context menu handler for
Directory, Directory\Background and Folders. But on Windows 7 (and likely
also Vista) this strategy results in duplication of the menu entries in
the explorer context menu. To deal with this a version field has been added
to specify the windows major version and the minimum minor version level
for which a registration entry should be active. So to only register
something on Windows XP the version field is set to 0x0500, while to
have an entry only register on Vista and Windows 7 set to 0x0600.

This has been tested on XP and Windows 7 but not on Vista.

To fix the build symbols for the versioning macros the build target is set
to Windows 2000 and above.

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>