allow debug strings larger than 1024 characters
On Windows vsnprintf returns -1 if the printed string is longer than the
target size. This is different from Linux where the number of characters
that would have been written if the target was large enough are returned.

Lets find out the target size first by printing into a string of zero
length. Interestingly this returns the number of characters to be written
on Windows.

This fixes a crash when the branchlist of repository is longer than 1024
characters.

Signed-off-by: Heiko Voigt <heiko.voigt@mahr.de>