use correct buffer size for vsnprintf in debug_git
This fixes a bug when debug_git is omitting some output that was
formatted by %i (or %x) at the end of the string.

It seems that when the final '\0' does not fit into the length given the
last format was omitted on mingw. Now the debug output works as
expected.

Signed-off-by: Heiko Voigt <hvoigt@hvoigt.net>