fix a crash when active_menu is NULL
On windows when right clicking on an item in a search result that was
moved to the trash the active_menu is NULL for some reason.

It is likely that build_cheetah_menu() has failed somewhere in the
callchain and thus left the active_menu invalid.

With this patch invoking the menu on such items does still not work but
the explorer does not crash anymore.

Signed-off-by: Heiko Voigt <hvoigt@hvoigt.net>