always read stdout and stderr from forked commands
When not doing so such output would be forwarded to a program using
this plugin. Normally in Windows GUI programs there is no console so
this did not disturb anyone.

Far Manager is a console program that also uses shell extensions for
which this was reported.

Fix this by always reading the output into a strbuf. Since we are
tracing the ouput later on anyway the output does not get lost.

Signed-off-by: Heiko Voigt <hvoigt@hvoigt.net>