Add a convenience goal 'checkinstall'
The checkinstall(8) tool is a very convenient tool to install packages
from source, integrating into the current system's default package
manager.

However, it does a slightly lousy job at determining package name
and version from the directory name 'git-cheetah': it tries to use
'git' as package name and 'cheetah' as version number (which fails
on Debian, because Debian requires version numbers to start with
a digit). So help it a little by adding a goal that automatically
determines correct default values.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>