Fix find_msysgit_in_path() when PATH is in cmd folder rather than bin
msysgit_in_path() expects the PATH to git.exe to be in "../bin/""
whereas Git for Windows uses "../cmd/". This results in the
PathToMsys registry value being set to "../cmd/git.exe" which means
that git_path() will always fail.

This only happens when the dll is outside the git-cheetah folder, as
can be the case when you are debugging.

Signed-off-by: johnstevenson <john-stevenson@blueyonder.co.uk>