exec.c: redirect unwanted command output to /dev/null
Git-Cheetah redirects and reads stdout / stderr of forked commands only if
it needs to. Otherwise this output simply goes to stdout / stderr of the
calling file manager process. This is OK for GUI-based file managers (e.g.
Windows Explorer), as there is no console.

For text-based file managers (e.g. Far Manager) this is a problem, as such
unwanted console output messes up the UI.

Redirect stdout / stderr of forked commands to /dev/null if Git-Cheetah
doesn't need it.

Signed-off-by: Karsten Blees <blees@dcon.de>