Detach Git Bash
When called from the Windows Explorer, we must not wait for the
stdout/stderr of Git Bash.

A recent change made for FarManager makes Git Cheetah capture
stderr/stdout even when we are not interested in it, to avoid cluttering
FarManager's precious console. Due to this workaround, Git Bash makes the
Explorer -- Git Cheetah's primary intended consumer -- hang.

Since Git Bash does not output anything to the calling console, let's just
force the detached mode by avoiding the FarManager workaround *just* for
Git Bash.

Thanks to John Stevenson for catching a bug in an earlier version of this
commit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>